+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- PrometheusIngester: added `query_type` label for `slo_exporter_prometheus_ingester_query_fails_total` metric
 
 ## [5.5.0] 2020-05-28
 ### Added